Move generators api into `generator.api` package
