+++ assertions/src/commonTest/kotlin/com/github/jcornaz/kwik/assertions/ForAll1Test.kt
@@ -1,7 +1,7 @@
 package com.github.jcornaz.kwik.assertions
 
-import com.github.jcornaz.kwik.generator.Generator
-import com.github.jcornaz.kwik.generator.withSamples
+import com.github.jcornaz.kwik.generator.api.Generator
+import com.github.jcornaz.kwik.generator.api.withSamples
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
