+++ generators-stdlib/src/commonMain/kotlin/com/github/jcornaz/kwik/generator/stdlib/Collections.kt
@@ -1,6 +1,6 @@
 package com.github.jcornaz.kwik.generator.stdlib
 
-import com.github.jcornaz.kwik.generator.Generator
+import com.github.jcornaz.kwik.generator.api.Generator
 import kotlin.random.Random
 
 private const val MAX_EXTRA_ADD_ATTEMPT = 1000
