+++ generators-stdlib/src/commonMain/kotlin/com/github/jcornaz/kwik/generator/stdlib/Defaults.kt
@@ -1,7 +1,7 @@
 package com.github.jcornaz.kwik.generator.stdlib
 
-import com.github.jcornaz.kwik.generator.*
-import com.github.jcornaz.kwik.generator.withNull
+import com.github.jcornaz.kwik.generator.api.Generator
+import com.github.jcornaz.kwik.generator.api.withNull
 
 /**
  * Returns default [Generator] for the type [T].
