+++ generators-stdlib/src/commonMain/kotlin/com/github/jcornaz/kwik/generator/stdlib/Enums.kt
@@ -1,6 +1,6 @@
 package com.github.jcornaz.kwik.generator.stdlib
 
-import com.github.jcornaz.kwik.generator.Generator
+import com.github.jcornaz.kwik.generator.api.Generator
 
 /**
  * Returns a [Generator] emitting random enum instance picked in the enum [T]
