+++ generators-test/src/commonMain/kotlin/com/github/jcornaz/kwik/generator/test/AbstractGeneratorTest.kt
@@ -1,6 +1,6 @@
 package com.github.jcornaz.kwik.generator.test
 
-import com.github.jcornaz.kwik.generator.Generator
+import com.github.jcornaz.kwik.generator.api.Generator
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNotEquals
